rename format-metrics-tables to format-tables
