+++ web/api/helpers/selfie-check-analytics/format-tables.ts
